Address detekt warnings after plugin changes
